FIX: setup dropdown on staking page
