+++ launcher/src/components/UI/staking-page/components/header/StakingHeader.vue
@@ -10,7 +10,7 @@
     </div>
     <div class="col-start-18 col-span-full h-8 flex justify-end items-center">
       <SetupDropdown
-        :list="setupsList"
+        :list="setupList"
         :new-height="newHeight"
         @select-setup="selectSetup"
         @server-view="serverView"
@@ -25,14 +25,36 @@ import SetupDropdown from "../../../edit-page/components/edit/setups/SetupDropdo
 import { useMultiSetups } from "@/composables/multiSetups";
 import TotalBalance from "../management/components/val-rewards/TotalBalance.vue";
 import NetworkStatus from "../../../../layers/NetworkStatus.vue";
+import { useServices } from "@/store/services";
+import { useDeepClone } from "@/composables/utils";
 
 const setupStore = useSetups();
+const serviceStore = useServices();
 const { getSelectedSetup, getServerView } = useMultiSetups();
 
 const newHeight = "h-8";
 
-const setupsList = computed(() => {
-  return setupStore.allSetups;
+const setupList = computed(() => {
+  const validators = serviceStore.installedServices
+    .filter((s) => s.category === "validator")
+    .map((v) => v.service);
+  let output = setupStore.allSetups.filter((s) => s.setupName !== "commonServices");
+
+  output = output.map((setup) => {
+    if (!setup.services || setup.services.length === 0) {
+      setup.noValidator = true;
+    } else {
+      const hasValidator = setup.services.some((service) =>
+        validators.includes(service.service)
+      );
+      if (!hasValidator) {
+        setup.noValidator = true;
+      }
+    }
+    return useDeepClone(setup);
+  });
+
+  return output;
 });
 
 const getNetworkSize = computed(() => {
